Allow multiple content souces
The page can now load the main content from either a .php or .html file
in the /pages directory, or from index.html or index.php in subdirectory
of the /pages directory. In all cases file or subdirectory name is to be
the page name used in the links created for navigation later.

Page content locations for `contact` would be, in order of search:

1. `/pages/contact.php`
2. `/pages/contact.html`
3. `/pages/contact/index.php`
4. `/pages/contact/index.html`

Includes an update to the `README` for typo fixes and status changes.