Convert config data into a class
Refactoring will lead to using functions that will not have the global
configuration variables in scope without using the `global` keyword,
which is considered "bad form". Using a class it can be accessable in
a controlled manner, as well as being unchangeable by the system after
the script is launched.